pyproject.toml
[tool.poetry]
name = "formed"
version = "0.0.1"
description = "⚙️ Flexible framework for organizing data/experiments/workflows"
authors = ["altescy <me@altescy.jp>"]
license = "MIT License"
readme = "README.md"
homepage = "https://altescy.jp/formed"
keywords=["python", "workflow"]

[tool.poetry.scripts]
formed = "formed.__main__:run"

[[tool.poetry.source]]
name = "pypi"
priority = "primary"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
colt = ">=0.16.1"
dill = "*"
filelock = "*"
pydantic = ">=2.0,<3.0"
pyyaml = "*"
rjsonnet = "*"
mlflow = {version = ">=2.0,<3.0", optional = true}
numpy = { version = ">=1.26.4", optional = true}
scipy = [
  {version = ">=1.14", python = ">=3.12", optional = true},
  {version = ">=1.13", python = ">=3.9,<3.12", optional = true},
]
pyarrow = [
  {version = ">=18.0", python = ">=3.12", optional = true},
  {version = "*", python = ">=3.9,<3.12", optional = true},
]

[tool.poetry.extras]
mlflow = ["mlflow", "numpy", "scipy", "pyarrow"]
all = ["mlflow", "numpy", "scipy", "pyarrow"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pysen = "^0.11.0"
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
mypy = "^1.11.2"
types-pyyaml = "^6.0.12.20240917"

[tool.poetry.group.lsp]
optional = true

[tool.poetry.group.lsp.dependencies]
python-lsp-server = { version="^1.12.0", extras=["flake8", "pylint"] }
pylsp-mypy = "^0.6.9"
python-lsp-isort = "^0.2.0"
python-lsp-black = "^2.0.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.5.49"
mkdocstrings = {version = "^0.27.0", extras = ["python"]}
mkdocs-autorefs = "^1.2.0"
mkdocs-material-extensions = "^1.3.1"
mkdocs = "^1.6.1"
pymdown-extensions = "^10.13"

[tool.pysen]
version = "0.11"

[tool.pysen-cli]
settings_dir = "."

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 120
py_version = "py39"
[[tool.pysen.lint.mypy_targets]]
  paths = ["."]

[tool.pysen.lint.source]
  includes = ["."]
  excludes = [".venv/"]


[tool.black] # automatically generated by pysen
# pysen ignores and overwrites any modifications
line-length = 120
target-version = ["py39"]

[tool.isort] # automatically generated by pysen
# pysen ignores and overwrites any modifications
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
line_length = 120
multi_line_output = 3
use_parentheses = true
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
